Added files via upload
Files include the Classes, Functions, and Scripts, but no results. Additionally, note that the addpath and genpath and load commands at the start of some scripts use a directory familiar to the computer that they were uploaded too, but may be incompatible with a general file structure. Correct this in a later commit.